pyproject.toml
[tool.poetry]
name = "cherry-in-the-haystack"
version = "0.0.1"
description = "AI Engineering knowledge handbook — curated, synthesized, always fresh"
authors = []
readme = "README.md"
packages = [{ include = "handbook" }]

[tool.poetry.dependencies]
python = "^3.10"
anthropic = ">=0.40"
notion-client = ">=2.2"
psycopg = { extras = ["binary", "pool"], version = ">=3.1" }
SPARQLWrapper = ">=2.0"
feedparser = ">=6.0"
tweepy = ">=4.14"
requests = ">=2.31"
python-dotenv = ">=1.0"
loguru = ">=0.7"
pymupdf = ">=1.24"
fastapi = ">=0.104"
uvicorn = { extras = ["standard"], version = ">=0.24" }

[tool.poetry.group.dev.dependencies]
ruff = ">=0.6"
mypy = ">=1.10"
pytest = ">=8.0"
pytest-cov = ">=5.0"
pytest-mock = ">=3.14"
pytest-asyncio = ">=0.23"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that require a live PostgreSQL instance (deselect with -m 'not integration')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
ignore_missing_imports = true
python_version = "3.10"
